More fixes for ckeditor when symfony is not running in root.
Remove leading / from output in javascript assetic block

The leading / was causing an issue when Symfony is not installed in root
(the subdirectory was not rendered in the script src).

Remove call to asset for ckeditor_basepath

PR #105 introduced an issue when the assets_version is set (the asset twig extension
appends the framework.templating.assets_version which breaks the path to ckeditor).
Also reset the default base path setting to include the leading / to be consistent
with the documentation.